+++ cmd/tlgen/typelang/parser.go
@@ -1,4 +1,4 @@
-package tlparser
+package typelang
 
 import (
 	"errors"
