+++ cmd/tlgen/typelang/parser_test.go
@@ -1,4 +1,4 @@
-package tlparser
+package typelang
 
 import (
 	"io/ioutil"
